+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - update base `DeltaLake` version to `0.7.0`.
 - update to Arc 3.0.0.
+- add `partitionBy` and `numPartitions` to `DeltaLakeMergeLoad`.
 
 ## 1.9.0